Implement BeginOverwritingVirtualUTXOSet and FinishOverwritingVirtualUTXOSet.
